experiment: allow custom http responses for non-swr requests
This experiment enables defining a custom HTTP response resolution
for requests that aren't made by the `use...` hooks created by the
library.

This allows people to re-use the logic in forms, which mean that people
can have progressive enhancement in their forms: if JS is not loaded,
you can submit the form "as is", and once JS is loaded, you will fire an
SWR mutation request. I think it's cool.

The only thing I'm not sure about is the API. This commit enables the
following API:

```ts
export const useMyMutation = mutation(parser, async () => output, {
  async resolveFormSubmission(output) {
    return new Response(...)
  }
})
```

This makes sense, but:

1. not every direct request is a form submission
2. how do we create form tags out of this?

We can expose a `Form` component that will receive the mutation/query.
But, that will mean we're going to have to have some metadata exposed on
the hook, like `useMyMutation.baseUrl` and `useMyMutation.method` (even
that we only allow `POST` right now for mutations and `GET` for queries)

This is kind of a smell but maybe it's okay because we already have this
data when we do the query 🤔 so we might extract it and reuse the same
value.

Maybe even not a `Form` but more of a `useFormSubmitCallback` that will
generate the necessary callback for forms?